Create Methodology Process.md
I believe on standing on the shoulders of giants.  I've provided the three seed links that I am starting out with on this, and then will adjust based on my experiences with other studying courses, CTFs, Cyber Ranges, etc. The goal is to make this a living document which will be tailored to my own style, and that you can tailor to your style in the future.